requirements.txt
alembic==1.8.1
anyio==3.6.1
black==22.8.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
cryptography==38.0.1
fastapi==0.81.0
FastAPI-SQLAlchemy==0.2.1
GeoAlchemy==0.7.2
GeoAlchemy2==0.12.5
greenlet==1.1.3
h11==0.13.0
httplib2==0.20.4
idna==3.3
Mako==1.2.2
mangum==0.15.1
MarkupSafe==2.1.1
mypy-extensions==0.4.3
numpy==1.23.2
oauth2client==4.1.3
packaging==21.3
pandas==1.4.4
pathspec==0.10.1
platformdirs==2.5.2
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.1
PyJWT==2.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
requests==2.28.1
rsa==4.9
sentry-sdk==1.9.7
Shapely==1.8.4
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.19.1
tomli==2.0.1
typing_extensions==4.3.0
urllib3==1.26.12
uvicorn==0.18.3
